Add shortcut restrictions & improving dismissal logic
